Add Edit function and rename modules
